EnhancedPDFLoader to PDFLoader
