+++ ingestion/document_processor.py
@@ -22,7 +22,7 @@ from core.graph_db import graph_db
 from core.ocr import ocr_processor
 from ingestion.loaders.csv_loader import CSVLoader
 from ingestion.loaders.docx_loader import DOCXLoader
-from ingestion.loaders.enhanced_pdf_loader import EnhancedPDFLoader
+from ingestion.loaders.pdf_loader import PDFLoader
 
 from ingestion.loaders.pptx_loader import PPTXLoader
 from ingestion.loaders.text_loader import TextLoader
@@ -64,7 +64,7 @@ class DocumentProcessor:
         """Initialize the document processor."""
         # Initialize loaders with OCR support
         self.loaders = {
-            ".pdf": EnhancedPDFLoader(),  # Use enhanced PDF loader with OCR
+            ".pdf": PDFLoader(),  # Use PDF loader with OCR
             ".docx": DOCXLoader(),
             ".txt": TextLoader(),
             ".md": TextLoader(),
